Support "px:extends" in script htmlization
